pyproject.toml
[tool.poetry]
name = "django_spaced_repetition_app"
version = "0.1.0"
description = ""
authors = ["vladislv31 <vladislv.k31@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
django-widget-tweaks = "^1.4.12"
django-debug-toolbar = "^3.6.0"
django-mptt = "^0.13.4"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
